requirements.txt
Jinja2==3.1.2
marko==1.2.1
MarkupSafe==2.1.1
python-frontmatter==1.0.0
PyYAML==6.0
